+++ lib/esdoc-plugin/doc.js
