+++ lib/esdoc-plugin/factory.js
@@ -1,5 +1,5 @@
 const assert = require('assert');
-const TestDoc = require('./TestDoc');
+const TestDoc = require('./doc');
 // hack: depends on ESDoc internal class
 const CommentParser = require('esdoc/out/src/Parser/CommentParser').default;
