+++ lib/esdoc-plugin/index.js
@@ -1,7 +1,7 @@
 const assert = require('assert');
 const path = require('path');
 const fs = require('fs');
-const TestDocFactory = require('./TestDocFactory');
+const TestDocFactory = require('./factory');
 
 // hack
 const ESParser = require('esdoc/out/src/Parser/ESParser').default;
